backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg[binary]>=3.2.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-socketio==5.11.0
aiofiles==23.2.1
python-dotenv==1.0.0
openai==1.10.0
Pillow==10.4.0
opencv-python==4.9.0.80
numpy==1.26.3
boto3==1.34.34
websockets==12.0
